sepolicy: Allow QFP daemon to send input events
The fingerprint daemon has a use-case in which the fingerprint
sensor is used as a home button (instead of a capacitive button).
When the user's finger touches the sensor, a "home button down"
event is sent to an input device.

In order to allow the daemon to send these events, we allow the
daemon to list the input devices in /dev/input and to write to
the char devices in /dev/input.

Change-Id: I1af95dcc5c72a3394e988b7dc23e7e1fc8d27a3b